Make OPTIONS responses cacheable (#40)
* Make OPTIONS responses cacheable

Should significantly reduce the number of requests to this service.  At the moment every time events are sent, an OPTIONS request is sent too. First attempt set the `cache-control` header but the browser seems to ignore that. Interestingly cloudfront is also ignoring the cache control headers, but since we aren't running a fully fledged cloudfront service (just the default one with API gateway), I don't think we can customise this behaviour. So I've removed the cache-control headers.